Revert "Adjust overflow handling and comment"
This reverts commit ceac7654aeb7ce4b2b91f6d84cf891c4bcdce755.

+++ src/libraries/System.Runtime.Numerics/src/System/Numerics/BigInteger.cs
@@ -1611,12 +1611,13 @@ namespace System.Numerics
                 // Let m=n*Log10(2), the final result would be x = (k*10^(m-[m]))*10^(i+[m])
 
                 ulong highBits = ((ulong)_bits[^1] << kcbitUint) + _bits[^2];
-                int lowBitsCount32 = _bits.Length - 2; // max length of BitInteger is Array.MaxLength/sizeof(uint), which is Array.MaxLength bits
+                double lowBitsCount32 = _bits.Length - 2; // if Length > int.MaxValue/32, counting in bits can cause overflow
                 double exponentLow = lowBitsCount32 * kcbitUint * Math.Log10(2);
 
-                // Max possible value < 2^Array.MaxLength < 2^int.MaxValue < 10^int.MaxValue
-                // No worrying about int overflow
-                int exponent = (int)exponentLow;
+                // Max possible length of _bits is int.MaxValue of bytes,
+                // thus max possible value of BigInteger is 2^(8*Array.MaxLength)-1 which is larger than 10^(2^33)
+                // Use long to avoid potential overflow
+                long exponent = (long)exponentLow;
                 double significand = (double)highBits * Math.Pow(10, exponentLow - exponent);
 
                 while (significand >= 10.0)
